DV-59 rename: 파일명에 공백 제거
