+++ src/components/interview/step/type-step.tsx
